git-svn-id: svn://svn.icms.temple.edu/lammps-ro/trunk@6530 f3b2605a-c512-4ea7-a41b-209d697bcdaa

+++ src/dump_image.h
@@ -60,8 +60,8 @@ class DumpImage : public DumpCustom {
   double ssaoint;
 
   int npixels,viewflag;
-  float *depthBuffer,*surfaceBuffer;
-  float *depthcopy,*surfacecopy;
+  double *depthBuffer,*surfaceBuffer;
+  double *depthcopy,*surfacecopy;
   char *imageBuffer,*rgbcopy,*writeBuffer;
 
   double **bufcopy;
@@ -89,9 +89,9 @@ class DumpImage : public DumpCustom {
   double specularHardness;
   double specularIntensity;
 
-  float SSAORadius;
+  double SSAORadius;
   int SSAOSamples;
-  float SSAOJitter;
+  double SSAOJitter;
 
   // dynamic view params
 
@@ -148,7 +148,7 @@ class DumpImage : public DumpCustom {
 
   void draw_sphere(double *, double *, double);
   void draw_cylinder(double *, double *, double *, double, int);
-  void draw_pixel(int, int, float, double *, double*);
+  void draw_pixel(int, int, double, double *, double*);
   void compute_SSAO();
   void write_JPG();
   void write_PPM();
